requirements.txt
# Core ML and Data Science
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
catboost==1.2.2

# Advanced ML and Optimization
optuna==3.4.0
hyperopt==0.2.7
imbalanced-learn==0.11.0
shap==0.43.0
lime==0.2.0.1

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
plotly-dash==2.16.1

# EDA and Profiling
pandas-profiling==3.6.6
sweetviz==2.3.1
ydata-profiling==4.6.4

# Web Framework and API
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
pydantic==2.5.2
python-multipart==0.0.6

# Data Processing and Feature Engineering
category-encoders==2.6.3
feature-engine==1.6.2
scipy==1.11.4

# Model Monitoring and MLOps
mlflow==2.8.1
evidently==0.4.11
great-expectations==0.18.5

# Testing and Code Quality
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
joblib==1.3.2
tqdm==4.66.1
requests==2.31.0

# Jupyter and Notebooks
jupyter==1.0.0
ipykernel==6.26.0
nbconvert==7.12.0

# Database and Storage
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Containerization support
gunicorn==21.2.0
